Raise an error from the Rspec::Parser if there are errors
RSpec exposes a 'summary' object, with counts of various things - in
this case, if the count of "errors_outside_of_examples" is above zero,
that's _different_ from a failing test, and we shouldn't let the
pipeline proceed.

Note that this wouldn't be an issue if rspec had chosen a different exit
status for "shit's broke" than for "some of your tests failed", but I
can understand why they decided not to do that - their test-runner
_itself_ isn't failing, just the test files it loads. (If the runner
failed, the parser would raise an error already, since it uses `fetch`
to get the 'examples' key, and JSON.parse on the stream).