chore: factor base_value slightly
This is the sort of method where the complexity checks are a little
annoying. A method lookup-table constant would make this needlessly
complex, but if we added another type of token we'd need to. (or more
probably, extract a domain object that decorates the message for this
purpose, and just `public_send` the parsed_token.source to it)