fix: handle the updated simplecov coverage messaging
When you enable simplecov, it writes plain text to stdout, *even if
rspec is running with `--format json`*. We have regexes to skip over
this plaintext, but they _updated_ it at some point, and our regexes
stopped matching.

Improve the regex to match both styles. Ideally, we'd use a json-parser
that knows how to skip all the junk at the end, but I haven't found one
yet (we'll adopt it if I do)